accept-terms split in 3 components
