+++ components/accept-terms-payment.jsx
@@ -6,25 +6,13 @@ import { AcceptTermsAccess } from './accept-terms-access';
 const DEFAULT_AGE_RESTRICTION = '16';
 const DEFAULT_PRIVACY_POLICIES_POSITION = 'top';
 
-export function AcceptTerms({
+export function AcceptTermsPayments({
 	withPrivacyPolicyTerms = false,
 	hasError = false,
 	isSignup = false,
-	isChecked = false,
-	isB2b = false,
-	isB2cPartnership = false,
-	ageRestriction = DEFAULT_AGE_RESTRICTION,
 	isEmbedded = false,
-	isCorpSignup = false,
 	isTrial = false,
-	isTransition = false,
-	transitionType = null,
 	isPrintProduct = false,
-	isSingleTerm = false,
-	isDeferredBilling = false,
-	hideConfirmTermsAndConditions = false,
-	children,
-	privacyPoliciesPosition = DEFAULT_PRIVACY_POLICIES_POSITION,
 }) {
 	const divProps = {
 		id: 'acceptTermsField',
@@ -32,138 +20,6 @@ export function AcceptTerms({
 		'data-validate': 'required,checked',
 		...(isSignup && { 'data-trackable': 'sign-up-terms' }),
 	};
-
-	const labelClassName = classNames([
-		'o-forms-input',
-		'o-forms-input--checkbox',
-		{
-			'o-forms-input--invalid': hasError,
-		},
-	]);
-
-	const inputProps = {
-		id: 'termsAcceptance',
-		type: 'checkbox',
-		name: 'termsAcceptance',
-		value: 'true',
-		'data-trackable': 'field-terms',
-		'aria-required': 'true',
-		required: true,
-		...(isChecked && { defaultChecked: true }),
-	};
-	/* 
-	const privacyPolicyTerms = (
-		<div className="privacy-policy-terms">
-			<span className={`consent-text--${privacyPoliciesPosition}`}>
-				For more information about how we use your data, please refer to our{' '}
-				<a
-					className="ncf__link--external"
-					href="https://help.ft.com/legal-privacy/privacy-policy/"
-					target="_blank"
-					rel="noopener noreferrer"
-					data-trackable="privacy-policy-info"
-				>
-					privacy
-				</a>{' '}
-				and&nbsp;
-				<a
-					className="ncf__link--external"
-					href="https://help.ft.com/legal-privacy/cookies/"
-					target="_blank"
-					rel="noopener noreferrer"
-					data-trackable="cookies-info"
-				>
-					cookie
-				</a>{' '}
-				policies.
-			</span>
-			{!hideConfirmTermsAndConditions && (
-				<label className={labelClassName} htmlFor="termsAcceptance">
-					<input {...inputProps} />
-					<span className="o-forms-input__label">
-						TESTEOI confirm that I am {ageRestriction} years or older and agree to the
-						full{' '}
-						<a
-							className="ncf__link--external"
-							href="http://help.ft.com/help/legal-privacy/terms-conditions/"
-							target={isEmbedded ? '_top' : '_blank'}
-							rel="noopener noreferrer"
-							data-trackable="terms-and-conditions"
-						>
-							Terms &amp; Conditions
-						</a>
-						.
-					</span>
-					<p className="o-forms-input__error">
-						Please accept our terms &amp; conditions
-					</p>
-				</label>
-			)}
-			{children && <div className="children-container">{children}</div>}
-		</div>
-	); */
-
-	const b2bTerms = isB2b ? (
-		<li>
-			<span className="terms-b2b">
-				By submitting this form, you indicate your consent to also being
-				contacted by Financial Times by email, post, or phone about our other
-				products, services or special offers unless you untick this box.
-			</span>
-		</li>
-	) : (
-		<li>
-			<span className="terms-default">
-				I confirm I am {ageRestriction} years or older and have read and agree
-				to the{' '}
-				<a
-					className="ncf__link--external"
-					href="http://help.ft.com/help/legal-privacy/terms-conditions/"
-					target={isEmbedded ? '_top' : '_blank'}
-					rel="noopener noreferrer"
-					data-trackable="terms-and-conditions"
-				>
-					Terms &amp; Conditions
-				</a>
-				.
-			</span>
-		</li>
-	);
-
-	const corpSignupTerms = isCorpSignup && (
-		<>
-			<li>
-				<span className="terms-corp-signup">
-					Your organisation’s administrator(s) may view basic usage and profile
-					data about your account and have the ability to set up myFT topic
-					follows on your behalf.
-				</span>
-			</li>
-			<li>
-				<span className="terms-corp-signup">
-					Basic usage and profile data about your account can include; for
-					example, your job title and profile information, the date you last
-					visited, volume of content consumed, etc.
-				</span>
-			</li>
-			<li>
-				<span className="terms-corp-signup">
-					myFT topics may be selected on your behalf by your company
-					administrator or FT representative for you to follow. You can unfollow
-					these topics or unsubscribe from the myFT digest through the Contact
-					preferences section on myFT.
-				</span>
-			</li>
-			{isTrial && (
-				<li>
-					<span className="terms-corp-signup">
-						This trial is to demonstrate the value of a group subscription and
-						we’ll contact you during your trial.
-					</span>
-				</li>
-			)}
-		</>
-	);
 
 	const transitionTerms = isTransition && (
 		<>
@@ -225,6 +81,15 @@ export function AcceptTerms({
 				</span>
 			</li>
 		</>
+	);
+
+	const deferredBillingTerms = isDeferredBilling && (
+		<li>
+			<span className="terms-deferred">
+				Please note if you fail to make payment for your deferred billing plan
+				within due date your subscription will be automatically cancelled.
+			</span>
+		</li>
 	);
 
 	const printSignupTermText = isTrial
@@ -301,46 +166,17 @@ export function AcceptTerms({
 		</>
 	);
 
-	const deferredBillingTerms = isDeferredBilling && (
-		<li>
-			<span className="terms-deferred">
-				Please note if you fail to make payment for your deferred billing plan
-				within due date your subscription will be automatically cancelled.
-			</span>
-		</li>
-	);
-
-	const b2cPartnershipTerms = (
-		<label className={[labelClassName, 'checkbox-two-lines'].join(' ')}>
-			<input {...inputProps} />
-			<span className="o-forms-input__label">
-				I confirm I am {ageRestriction} years or older and have read and agree
-				to the{' '}
-				<a
-					className="ncf__link--external"
-					href="http://help.ft.com/help/legal-privacy/terms-conditions/"
-					target={isEmbedded ? '_top' : '_blank'}
-					rel="noopener noreferrer"
-					data-trackable="terms-and-conditions"
-				>
-					Terms &amp; Conditions
-				</a>
-				.
-			</span>
-			<p className="o-forms-input__error">
-				Please accept our terms &amp; conditions
-			</p>
-		</label>
-	);
-
 	return (
 		<div {...divProps}>
-			{isB2cPartnership ? b2cPartnershipTerms : <AcceptTermsAccess />}
+			{signupTerms}
+			{/* transitionTerms and deferredBillingTerms used in next-retention, condition to render? */}
+			{transitionTerms}
+			{deferredBillingTerms}
 		</div>
 	);
 }
 
-AcceptTerms.propTypes = {
+AcceptTermsPayments.propTypes = {
 	withPrivacyPolicyTerms: PropTypes.bool,
 	hasError: PropTypes.bool,
 	isSignup: PropTypes.bool,
